Refactor backend: split monolithic `app.ts` into routes/controllers/schemas/logging Fixes #82
