+++ memory/tasks/_index.md
@@ -24,6 +24,6 @@
 - TASK022 - Copilot CLI model probe (/models endpoint) - Completed ✅ (2026-01-25)
 - TASK023 - Shared RedVsBlue config (DES022) - Completed ✅ (2026-01-25)
 - TASK024 - RedVsBlue match store split - Completed ✅ (2026-01-26)
-- TASK025 - Refactor backend: split `app.ts` into routes/controllers/schemas/logging - Pending (2026-01-26)
+- TASK025 - Refactor backend: split `app.ts` into routes/controllers/schemas/logging - Completed ✅ (2026-01-26)
 - TASK026 - Extract engine & renderer responsibilities from `useGame` - Pending (2026-01-26)
 - TASK027 - Decompose `RedVsBlue.tsx` into match/session, AI director, and toast hooks - Pending (2026-01-26)
